Fix secret creation when pod created via deployment
When a deployment triggers pod creation, the admissionreview
request does not appear to have a fully formed Pod for its Object
field.
Updated the admission logic to respond to the creation of
deployments, statefulsets, replicasets and daemonsets, creating
the secret in the target namespace when one is encountered.
As a corollory to this, restricted the mutation logic to add
imagePullSecrets to Pod kinds exclusively