Warn when secret creation fails
If secret creation in a namespace fails we fail the http request.
This now warns, but otherwise continues.

This caters for the case where secret creation fails because the
secret has been created already, but an additional retry mechanism
should be added to cater for other transient issues with credential
creation