feat: always update regcred secret with latest values
If the regcred secret exists in a namespace, update it with the
latest values. This avoids having to delete old secrets when
cycling the docker secrets

Added a random annotation to the deployment pods to force
re-deployment of the pods on every change (e.g. when secrets /
TLS certs change)